Add setuptools, PYTARGETS rule for better python support.
This changes allows any python package the uses the standard setup
mechanism to install to call PYTARGETS rather than GENTARGETS in the
packages makefile.  This does require unicodedata and zlib support in
the host build, so these are un-disabled now.  It also requires zlib
support on the target, so it is enabled there as well.