Add toolbox-for-linux package to buildroot.
Toolbox-for-linux is a BSD-licensed replacement for busybox.  This
commit adds the target to buildroot and disables busybox if toolbox is
enabled; currently it is NOT enabled.  Replacing busybox with toolbox
requires the coordination of other packages (dhcp, coreutils, ntp,
bridge to start with) and is not happening in this commit.

Change-Id: Ife0aaaa4af2c1b8c06280c554d1418c93b2c91ca