configs/bruno*: use -j15 instead of -j6
Our workstations have lots of RAM and lots of threads available, but most of
the time spent building is spent sitting around with a load average <= 2.
Let's try (a bit feebly) to let buildroot do more things at once.

(The core problem is it appears to only want to build a single package at a
time; fixing that is much harder.)

Change-Id: I9042376a23fcf622a4959919cd6cd1f6be7eebfc