Enable BR2_PACKAGE_GOOGLE_VMLINUZ for the initramfs build.
Also enable the UBI/UBIFS config options, which it depends on so it can grab
the config settings.  (This is kind of disappointing since we don't actually
want to build an UBI version of the rootfs in the initramfs build, we just
want the parameters.  But it's straightforward and works.)

And now the app/ build no longer needs GOOGLE_VMLINUZ, so disable it.  And
in builder.py, we can build the kernel UBI image without building app, so
remove the check for that option.

Change-Id: Icb697ebc1e2c8ba6f72c1d27a91f5819387c3f2e