sysklogd: only change output format for /dev/kmsg.
My previous patch changed the syslogd line output format for all outputs,
because it was originally intended to *only* write to /dev/kmsg.  But then
we brought back /var/log/kern.log, and that one changed format too, which is
undesirable.

Make the patch a little more complicated so it only changes the format of
output to /dev/kmsg.

Change-Id: I6fbb142269e51afa05cdf0b9634364feec453cbd