simpleramfs: get rid of the init/app split.
Instead of doing all the complicated stuff we were doing before, just
generate the initramfs by copying files programmatically from the main
squashfs using a simple script.  This simplifies lots of code, avoids
double-building a bunch of things, and shrinks the kernel+initramfs down to
about 5MB instead of 30+MB.  (Of course, it does this by removing a bunch of
functionality from initramfs.  Supposedly, functionality that we weren't
using anyway...)

Change-Id: I32309cae748cb6bc9c9d79244614f6965de2c882