New bruno version number / ginstall filename algorithm.
We now just produce a /etc/version containing either:
	$(git describe)
or
	$(git describe)-$initials
or
	$(git describe)-$initials-dirty

Where $initials is the first two letters of $(whoami).  If you build with
--production, we don't add the initials, otherwise we do.  If the tree has
changes that aren't yet committed, we append '-dirty'.

Then we name the produced ginstall image file like this:

	bruno-$version.gi

	(for example:
	  bruno-20120430-eng-10-g5400ccc-ap.gi
	  bruno-flamingo4.gi
	  bruno-flamingo4-10-g5400ccc-dg.gi)

and symlink that to the old bruno_ginstall_image.tgz.

The idea of the new naming is that "official" builds have less crufty names,
because something like 'flamingo4' is enough to uniquely identify a tagged
revision built by the official build account.  Versions built from
non-tagged versions, or by non-official accounts, have extra stuff tacked
on.  I changed the file extension since a) the tools themselves don't care,
and b) this avoids the need to mention the fact that it's an installer as
part of the filename proper.

Change-Id: I518e1b4a3e8fa325ce13cba46065e9b07839a260