simpleramfs+ginstall: include a ramfs file that we can substitute into.
We create /rootfs.sum in simpleramfs, which has two markers and a bunch of
blank space.  As part of ginstall.mk, we run a little script that replaces
the blank space with a string of your choosing (and pad it out so the binary
file container doesn't get confused).

In order to make this work, we have to stop compressing the initramfs cpio
file before putting it in the kernel (otherwise the marker strings wouldn't
be findable, and the block size would be variable).  This doesn't seem to
matter since we compress the full kernel afterwards anyway, so the size
remains about the same and in fact it probably boots microscopically faster
now since it doesn't have to double-decompress the ramfs.

Right now we just replace the filler with INSERT_ACTUAL_DATA_HERE.  That
actual data should be the checksum of the rootfs.  I expect that will come
in a separate patch that also verifies the checksum at boot time.

Change-Id: Id3c667581b83f28c77f9e89569efb33c755a6ec2