Put fs skeleton in package/bruno/skel/.
Moving it from vendor/google/platform/skel/.  The skel almost always changes
as part of changing the packages in buildroot, not so much when changing the
random tools in vendor/google/platform/, so it's better as part of this repo
instead.

Also, we had confusing behaviour before: we were copying
*both* fs/skeleton/ (once upon first build) *and* our own skel (every time
when finalizing the rootfs), which is why certain weird files
(like /etc/shadow) would end up appearing sometimes, but not other times.
This patch moves the necessary missing files from fs/skeleton/ and now *all*
the files are copied correctly while finalizing the rootfs.

When copying the skeleton dir, hardlink the files rather than copying them,
so that editing them in buildroot takes effect immediately in the target
directory (and vice versa).  This is handy for NFS root.

Don't just copy the skeleton dir once the very first time we build; have it
depend on all the files inside the skeleton source dir, so if they change,
we relink the files (and most importantly, we link any new ones).

Arguably we should just *replace* fs/skeleton instead of using our own
directory.  That would cause conflicts with later versions of buildroot
(easy to resolve by just rejecting their changes) but would also mean we
automatically get improvements from later versions of buildroot, like new
versions of /etc/services.  I'm not doing that for now, though I could go
either way.

While I was here, removed /etc/qos.* because we're not using it nowadays.

Change-Id: I65aa598fe43bd2eb220bf48f02a1ecc4af28f316