bcm_apps: try to get playready.bin from a static file.
If /usr/local/google/playready_dev_license exists, we won't try to do a
'blaze build' to produce it.  This makes it so autobuilder machines/accounts
don't need to have prodaccess.

It also looks for playready_prod_license when building a PROD image,
but we may not be allowed to leave that file lying around. Doesn't matter
too much right now since the autobuilder doesn't try to build PROD images.

Change-Id: Id70bb9c4bd816d26d21edfed43729c9531fb4db7