.root shouldn't depend on filesystem contents.
In 022bea4e6626f0aa54b98505468c8d5573c4cbd6 I changed the Makefile rule for
$O/.root to depend on the contents of the skeleton directory.  This seemed
clever, but the problem is, the .root stamp is depended-upon by lots of
other stuff (to ensure the rootfs skeleton is installed before trying to
install anything else into it) so this means any change to the skel/
directory will suddenly cause virtually *everything* to reconfigure and
rebuild, and some packages (notably python) die when you do that without
deleting them first.

That's actually a bug in the packages, but anyway, we certainly don't want
to rebuild everything from scratch so often, so let's remove this extra
dependency.

It turns out not to matter anyway: in .config we still run copy_skel.sh,
which (re)copies the skel directory just before building the rootdisk
anyway.  This was needed because otherwise, packages might overwrite some of
our stuff (eg. the /etc/init.d/ files we painstakingly customized) and we
want to put it back.  But the net effect is we don't need to re-run the
.root rule anyway.

Change-Id: Ib78b2b706b1a8e5f80f21039d4f55d2fe5c1c8ec