Package cfe in buildroot.
This package is added just for the conveniece of building different
varieties of cfe and automate the signing process for it. To build cfe,
as long as you have access to production key, you can do the following

make O=[output dir] bcm_cfe-dirclean bcm_cfe

All cfe images and signatures shall be present in [output dir]/images
directory.

Change-Id: Ic6dcbe9c1d5e97abb8d8661faa42ab143772f23f