Don't install bruno.pc; use $BRUNO_DEFINES instead.
bcm_nexus was already checking for a $BRUNO_DEFINES variable in its build
system, but it wasn't exported.  That's all bruno.pc was providing; for
simplicity, pass it along when building bcm_*, and get rid of bruno.pc.

(The main motivation here is to avoid needing to install google_platform_*
before compiling bcm_*.  The dependencies really should run the other way;
bruno depends on bcm.  bcm is just a library that shouldn't depend on our
platform.)

Change-Id: I6fdd51b98701db49e188e6e7a93ce2808f84482d