Add 'zap' command to do a factory reset.
This is pretty simple for now: we just wipe the 'user+ubi' and 'misc+ubi'
partitions.  That gets rid of all user-configured data.

I think SageTV might need a change to keep a copy of its media database on
the disk as well as the flash, since we're intentionally *not* wiping the
disk.  (Just because we want to reset the device doesn't mean we want to
lose all our recordings!)

Change-Id: I7ee6167484585b7c6be807b93b19ac8d93366ab7