Do a factory reset when the reset button is pressed.
This works on both fat and thin bruno.  If the reset button is pressed for
about 4 seconds after starting userspace (which you can see because the led
turns from red to flashing blue), we blink all the lights noisily and run
the 'zap' script, then reboot.

If the reset button is not pressed at boot, but *is* held for more than 4
seconds at runtime, we run 'ginstall -p other' to swap to the boot image
that's not currently running.  This is a last-resort tactic in case the
current image craps out for some reason.  Note that it's not foolproof,
because the current image is the one responsible for switching to the
alternate one.  But it could help if, say, the system mostly boots but TR-069
doesn't work.

Change-Id: Id27b1a884b5a4b2d2b9092bf2f4a3cdc7619d362