cpulog: call sys.stdout.flush().
This corrects a problem where the messages would only be dumped to stdout,
all at once, every x messages, all with the same timestamp.  With the
default delay of 600sec this would take a *long* time.

To see the problem, compare the output of these commands before and after:
	./cpulog -i1 -w1
	./cpulog -i1 -w1 | grep .

Also fixed some gpylint messages while we're here.

Change-Id: I7d2525839eede7d61fb8af5de27139e106d02488