cpulog: print times in fractional seconds instead of msec.
This isn't really essential, but it was bothering me that the time period
was in seconds, but the per-process usage was msec.  If we print the number
of seconds instead (to three decimal places) we get equal precision but
don't have to specify the unit each time.  Also there's less chance of a
human misreading the units.

Change-Id: I28cd9fe8a4c7f064cebdf5d5e6b3509422ffce17