stresstest: wait for other end's network to be ready.
There was a race during stresstest startup: we started the iperf client
right away, and if the server wasn't running, we'd just auto-restart it if
it died.  Since we have to start both servers and they both do this, there
was no way to start them both without needing at least one retry.  However,
the auto-restart logic hid the fact that sometimes we wouldn't get *any*
connection to the remote machine, so the stresstest wasn't actually
stressing the network.

Instead, let's put in an explicit loop waiting for the remote end to start
its servers before we actually run the real iperf client.  Then, once it's
started, if it exits with nonzero, that's a test failure, so we'll abort
right away.

(I would like to run iperf 'forever' and not loop it, but unfortunately it's
pretty dumb, and will happily freeze up forever and/or report 0 bytes/sec if
the other end disappears.  So we let it run for 60 seconds between restarts,
and hope connect() times out the next time if there's a problem.  To catch
this requires the iperf patch in the previous commit, so it returns nonzero
in such cases.)

Also, if the $DONT_ABORT environment variable is set, stresstest keeps
running forever instead of aborting when something fishy happens.  That's
not so good for a system stability test, but if you just want to melt the
hardware for the MTBF test, it's probably better this way.

Change-Id: Ic274ddd545f24abaca0d634dc287068190e0867e