/usr/bin/burnin-wifi-setup: wlan config scripts used by humax.
This is a variant of the wifi setup script used by humax for stress testing
setup.  It disconnects the wifi device from br0, then connects to the given
SSID and obtained an address from DHCP for testing.

Nothing runs this for you: you should run it once before starting
stresstest.

Change-Id: I172bb2d72c0a681b9c8e446e8afdfc7c5540cc86