loglinear: a script for sorting parallel make output.
This adds the loglinear script and makes basic changes to the buildroot
makefiles so that it runs loglinear in the right places.

This makes the make output more bursty (it doesn't print anything until an
entire submake has finished) but much more readable, because the output from
one process is never interspersed with the output from another.  Also, if
there's an error, non-errored process outputs are printed first, and the
innermost errored process outputs are printed last.  Hopefully this means
the "interesting" part of any error will be right at the bottom of the
output in most cases.

Also, each line is prefixed with a string describing the sequence of
submakes that ended up invoking that line.  This makes the text more
greppable, and when there's an error, it's easier to see which make target
caused it.

Change-Id: I2346adfc005f6cdd8eddb17c7a7b3bfc5e66b0ac