pkillwait: a script to SIGTERM, then SIGKILL.
We have a lot of scripts that try to pkill processes to kill them, and the
'restart' implementation just has a hardcoded 2-second sleep, which is
annoyingly long and yet sometimes too short.

Use this script instead to kill a program with SIGTERM, then if it's not
dead after two seconds, SIGKILL it.  But if it dies sooner, it returns
sooner, making a typical restart happen faster.

Change-Id: I8fdd6d6f21410f34abdf8dd3c89f10d07fd3fca0