dhclient-script: don't let wifi interface own default route.
In burnin-wifi-setup, we take eth2 out of br0 and let it have its own
dhclient.  But we don't want that dhclient to have a default gateway; we
always want to use the wired interface for that.

(This is mostly for the manufacturing tests at Humax.  I'm not totally happy
with this solution; the "right" thing would be for their wifi routers to not
hand out default gateways, if they don't have routes to the Internet
anyhow.  But that's hard to configure on some routers, and this is a bit
more foolproof.  On the other hand, I really hate hardcoding eth2==wifi and
specific routing-related policy like this.)

Change-Id: I6ccbe3d5491440e405296c6227ecd872601e6f1f