Add ubiscrubber to fix bit-flips in UBI volumes.
ubiscrubber successively reads all pages of a UBI volume to allow the
UBI layer to discover and fix bit-flips (ECC errors). The rate at which
pages are read is determined by the age of the flash device. The older
the flash device, the higher the rate at which scrubber operates. The
age is defined by the maximum erase count of all erase blocks.

Add /etc/init.d/S94ubiscrubber which discovers all UBI volumes in the
system and starts an instance of ubiscrubber for each volume.

Change-Id: I05b7769712c94ea35f7f3ea75bee746dd6fcd6ea