Revert "stresstest: pkillwait dhclient before continuing."
Now that we no longer hardcode IP addresses in the stresstest, there's no
need to kill the dhclient anymore.  In addition to being generally less
gross, this fixes a bug where bruno would get an address on the ipv6 network
before IPv4 would finish responding, causing the stresstest to not work
because it killed dhclient too soon.

This reverts commit 113c4d5471664c8e78abe12815f3263072fe9fe6.

Change-Id: I1eef23d6b8c68174cc579cf7aa7703caa076ca4e