miniclient: start waitpower at bootup; clean up scripts.
According to b/6734557, we want the system to boot in powered-off mode, so
that after a power failure, people's speakers don't start blaring in the
middle of the night when the power comes back.

While we're here, rewrite the whole runminiclient loop to make it more
understandable and have less duplication.

Change-Id: If4d93f2c9e2d1ee7e941d37ba138de4bed81d01d