procps: INSTALL targets should be .PHONY
Without this patch, if /bin/ps had a newer timestamp than the ps binary
built by this package, 'make install' would fail to install it.  That sounds
right, but it's not what people expect when they type 'make install'.  In
particular, if /bin/ps was provided by something like busybox, we really do
want to replace it with the procps version.

Change-Id: I3652026499c43b0a79e22237f1157e37d252323c