Hardcode rootfstype to squashfs.
Currently cfe read from EXTRA_KERNEL_OPT and populate cmdline with the
content in EXTRA_KERNEL_OPT. We don't really need EXTRA_KERNEL_OPT if we
can figure out the rootfs type.

To prepare for removing EXTRA_KERNEL_OPT from cfe, in mounts-root, we
don't read rootfstype any more, instead, rootfstype is defaulted to
squashfs which is the only one we use on nand flash.

In the future, if we introduce new rootfs type, we can either fix it in
mounts-root, or port the mount.c with file system detection.

Change-Id: Iac04d33363257ebe1de9735cbdadd8df8c1e1622