iperf: Option --priority sets SO_PRIORITY
Add command line option --priority that allows the user to specify the
protocol-defined priority i.e. the skb_priority which is the priority in
the socket buffer (sk_buff). Packets with a higher priority may be
processed first depending on the selected device queueing discipline.
Furthermore, this value eventually maps to the Priority Code Point (PCP)
in the 802.1q VLAN tag. The exact mapping can be found in the output of
the following command under "egress-qos-map".

ip -d link show dev eth1.0

Also, look at socket(7) and search for SO_PRIORITY.

Change-Id: I8e3d07afc2a47019d797fe58bc596116415fafc8