Add package/bruno/skel/usr/share/LICENSES.
This file was partly auto-generated using the 'make legal-info' feature in
later versions of buildroot (hackily merged into our tree so that it would
work).  Sadly it's missing a lot of packages, still, so I had to do some
hand rolling for those ones (as well as the obvious ones that are not part
of the upstream buildroot).  Eventually we'll upgrade buildroot and be able
to do more of it automatically.  But this should take care of our license
compliance issues for now.

The file is displayed in the SageTV UI when requested.

Change-Id: I0bf6df581cf2f319ba3dc7c9039e89a9a438f0a5