samba: disable symlink() and change permissions so unlink() is allowed.
I don't really know why samba requires you to have *both* group and user
write permissions on a containing directory in order to be allowed to delete
a file in that directory.  Maybe it's a bug, or maybe it has to do with my
earlier setuid patch.  In any case, since the owner is video.video, it's
find to set our directories to gruop writable.

Change-Id: Ia8d1522086496399f6c7f93c69993b7834d12e79