builder.py: remove -p, -m, -c options and replace with -C
We were auto-generating a defconfig filename by combining those three
options, but it was actually less flexible than a simple -C option to
literally provide the defconfig filename.  Doing it this way will make it
easier for building on completely different hardware platforms, or
specifying the bruno_toolchain defconfig, etc.

Change-Id: Ie62bc6634c18fc926e350c5de852997f255c1b31