Rearrange some .config options, and change loader.bin location.
Find loader.bin in vendor/broadcom/cfe-bin, not vendor/google/platform.  The
latter is being scrubbed of proprietary binaries.  Also, the loader was not
actually used by package/bruno, so remove all references to it from there;
just use it directly in fs/ginstall instead.

We add a new BR2_PACKAGE_BRUNO_OPENBOX setting for selecting the openbox
bootloader in a build, and a --openbox option to builder.py to build it
(which forces BR2_PACKAGE_BRUNO_PROD to off).

While we're here, move a few packages around in Config.in and remove some
obsolete config options.  We don't need BR2_BRUNO_BCHP_VER anymore because
we build unified binaries that run on all our platforms.  We don't need
BR2_PACKAGE_BRUNO_DOGFOOD because we've always just used a single build with
the dogfood binary enabled or disabled according to system configuration.

Also, packages needed only incidentally and not for building (like hdparm)
are moved to defconfig (rather than a dependency in
package/bruno/Config.in).  The rule of thumb is:

  - if it's included in openbox, put it in defconfig
  - if it's something extra needed for a full bruno build, put it in the
      BR2_PACKAGE_BRUNO_APPS dependencies.

This hopefully maintains the spirit of the .config file (people doing the
build can twiddle options if they want) while maintaining our ability to
easily build "app" vs. "basic platform" builds.

Change-Id: I74b5f8b6de8823a3df2b60da303756fbd75d6e06