Add a 'make depcheck' target and fix missing Config.in dependencies.
We had a bunch of correctly-declared dependencies in *.mk, but we weren't
also setting them in Config.in, which resulted in some BR2_PACKAGE_*
variables not being set, which could lead to problems.  The main problem I
observed was that 'make source' would not actually download/extract all the
source for all the packages we used, because it only uses dependency
information from *.mk.  So we'd sometimes end up trying to download packages
at build time instead of during the initial sanity check.

This was particularly annoying because it cluttered the 'finaltargets'
section of the make output, with lines that rarely failed but often
obfuscated the real cause of a build failure when it happened.

Change-Id: I7cd96208a9c947eed2d11e9b7d6922077515d2e5