Revert "The GET to dropbox is no longer required."
We need a way to get some subset of devices to send their logs
to the canary test instance.  This seems to be the easiest way,
the GET can just return the URL of the canary to POST to, and
the dropbox app can have some datastore data that holds the set
to redirect.  When we switch over to the borgifired version that
set can be in CDD.

This reverts commit 2cf22e07ebabdeea06f72aa6b197818e27c66f6b.

Change-Id: I00d7bdf46fea56fd070b1682f9bbdd3c151099eb