set-acs: near-total rewrite, plus unit tests.
It all began with the realization that acs_url.cwmp was not one thing, but
many things: any of the URLs (the one provided by DHCP, or manually, or
find-acs, or...) were equally likely to redirect you to another server.
And, as it turns out, it's *never* safe to discard that redirect
information, because the server that redirected you might be shut down
someday (which is what happened to our test environment).  Moreover, if
server A redirects to server B, there's nothing stopping server B from
redirecting onward to server C, and so on, and all those links in the chain
are important.

Also, if you have a redirect from server A to server B, but server A was
just the default hardcoded ACS URL, and a DHCP D server then appears, the
DHCP server should take priority.  On the other hand, if D then redirects to
E, E takes priority.  You can't just assign "redirects" at a single level of
precedence - they're tied to the thing they're redirecting from.

Of course, the fun doesn't stop there.  If your DHCP server gave you an URL
yesterday, but it doesn't today, what do you do?  Even better, what if it
gives you a lying URL that's just wrong?  All of these things have happened
to us.  This code makes good stuff happen in all those situations.

It's frightening, basically.  But at least it has tests now.

Note: requires slight changes to cwmpd to support the new 'set-acs bless'
operation.

Fixes b/8008584

Change-Id: I4fd548d6d5590280eb3d94321e0628cd1153cf6c