S60harddisk: mount ext4 with barrier=1 instead of barrier=0.
My benchmarks suggest that there is no measurable performance difference
between the two, for our use case.  (This makes sense since it probably only
affects performance by adding a seek or two during metadata operations, such
as fallocate(), which happens seldom.)

On the other hand, setting barrier=0 could effectively bypass the journal
and allow for filesystem corruption in the event of a power failure or
surprise reset (such as hardware watchdog or kernel panic), because it
depends on the disk's internal cache to get flushed at the right time.  This
is probably the cause of the (minor) ext4 corruption errors we see in the
logs for a small number of devices.  (Those are separate errors from
SATA read errors, which are probably hardware problems.)

We keep data=writeback, because that doesn't cause filesystem-level
corruption in the case of power failure.  (It could cause junk to get
inserted into the unwritten part of a fallocate()ed file, but that doesn't
really matter to us since the probability is low and SageTV knows how to
reject junk input anyway.)

Also, move the mount flags into a FLAGS variable because we had already
grown at least one if/else branch that used different flags than the others,
for no good reason.

Change-Id: Ic801cd54d88ecb31841d1edb63fde2654c16122e