Add freegrep package.
The minimalist grep we've been using is really bad, lacking things like
square brackets and -v.  We want to avoid GNU tools because they're bloated,
but freegrep (based on the OpenBSD grep source) is pretty small and
featureful, and requires only minor changes to make it work on Linux, so
let's use that.

Change-Id: I3bfe0988d9bf17ac7c6745e19a4fc97551f23bc3