h/w watchdog timer triggers NMI then chip reset
Previously, a h/w watchdog timeout triggered a chip reset after 80
seconds which resulted in loss of the dmesg ring buffer in the kernel.

Let's have the h/w watchdog trigger an NMI instead and develop an NMI
handler that dumps the CPU registers and performs a soft reboot. This
way, the dmesg ring buffer is preserved, and we are in a better position
to debug the problem. If, for some reason, the CPU does not react to the
NMI, the h/w watchdog will eventually assert a chip reset event and
allow the box to recover.

Change-Id: I98104cb376b85f054f9499962a5ae93de7b33bca