Allow /etc/init.d scripts to be skipped.
I added a new file called init_list in /config/ that contains the names
of the init scripts to be skipped.
/config/init_list should have the full path for the file we want to
skip:
e.g.: /etc/init.d/S30bsa
I also added a new function in /etc/utils.sh that checks if the desired
file name is found in a given file.

Change-Id: I7feb029b9a17e2bfe0123a86d244c63140a14066