Remove softdog (software watchdog).
Since danielmentz@ improved our hardware watchdog support to send NMIs and
print more useful debug messages, there seems to no longer be any remaining
reason to run a software watchdog.  Moreover, we don't print the same level
of useful traces when a softdog occurs, so at this point it's making things
worse instead of better.

There's a slight possibility that maybe the hardware watchdog will fail to
kick in in some rare circumstances that would have been caught by the
softdog, but I think that's fairly unlikely.  We should keep our eyes open
if it does happen.  In a lab it'll be obvious (device crashes and forgets to
reboot) but in the field less so (device always gets rebooted sooner or
later, if only by an angry owner).

Change-Id: I9ff3a51c56c02f40c4ada34ab79c5a72b10c9f16