babysit and watchdog: add OOM killer resilience.
The kernel OOM killer, when it decides to kill something, sends SIGKILL,
which would cause the babysitter to *not* restart it because it thought you
were doing it on purpose.  The main reason we needed that was because
pkillwait sends SIGKILL sometimes.  So instead, we change pkillwait to also
kill the babysitters relevant to the task you're killing, which we probably
should have done all along.

Also, under some rare conditions the OOM killer would kill the watchdog or
the babysitter tasks themselves.  This isn't helpful.  Instead, adjust the
OOM killer priority so that the watchdog never dies, and babysitters almost
never die (unless things are getting totally crazy).  That way if a task
gets wildly out of control, the kernel can kill it off and the babysitter
can start a fresh one.

Change-Id: I9a3007e086f5da878372ea6dbb4d45d1402d682c