Print a message when finished writing a core dump.
Right now, some processes are rebooting the kernel after a core dump with a
the "INFO: task xxxx:yyyy blocked for more than 120 seconds" message.  We
know the rest of the system is still working fine at that point, but we
don't know if the problem is that the coredump never finished, or that the
program stayed stuck even after the coredump finished.

Print a message when we're finished so we can tell the difference.

Change-Id: I4aa6cf6c57e1676a3796cda047b6108c7b7d5b63