Handle skeleton files *.platform_xyz correctly
The suffix .platform_xyz can be appended to file names in fs/skeleton to
specify the platform they are meant for (gfibertv or gfiberlt). If a
filename carries the .platform_xyz suffix, then it is copied to the
target directory iff xyz is the platform currently being built.

The previous approach was confusing: All files (even the ones for
different platforms) in fs/skeleton were copied to the target directory
with the .platform_xyz suffix still part of the filenames. Only later,
postbuild.sh would remove files with the .platform_xyz suffix from the
target directory and copy only the correct subset of files for the
platform being built and with the suffix removed to the target
directory.

Another issue with the previous approach was that the following files
were missing from the target directory:

/samba/lmhosts
/usr/lib/media/landscapes/*.jpg

Also, when executing make in the out/ directory after the build has
already finished, we ended up with the following superfluous directory:

/etc/samba/etc/etc.platform_gfibertv

Change-Id: I16697fc94ff7daa7891c3c96d3b40e4153a5710e