Remove diagd and sysklogd.
As far as I know, our only remaining use of syslogd/klogd was to support
diagd.  And diagd is obsolete nowadays thanks to cwmp periodicstatistics and
the log uploader.

Without diagd, the bruno package no longer cares about the moca package,
which means we can remove all dependency on the broadcom packages.

Change-Id: I62870a4513d78bc4438287a11ea9cd8922116117