kvm_defconfig: switch to using x86_64 instead of x86 32-bit build.
We also switch from uClibc to glibc for this build.  The advantage of these
changes is that the produced binaries can all run on a plain ubuntu
workstation.

Unfortunately, buildroot's normal toolchain compiler can't handle glibc, so
we had to use crosstool, which is not so bad, but adds a few layers of
complexity.  And glibc introduced a couple of compiler problems which we
workaround with a couple of minor package patches.

Change-Id: Ic262e473b0733062624ff12040aa8fb492fe706f