gftv100_defconfig: re-enable ACL library.
I don't think we actually need ACLs, but unfortunately turning it off in one
branch but not another triggers some kind of weird ccache bug that prevents
samba from building.  The fault seems to be in ccache (clearing the cache
fixes it) but I really don't have time to try to debug that right now, and
the space savings isn't that significant.  This fixes the autobuilder.

Change-Id: I0c00688c42e038fe406920cf434d648ca3cdeed9