pkillwait: *actually* kill babysitters.
An earlier attempt to make pkillwait kill babysitters as well as babysat
processes didn't quite work out.  It would wait for the babysitter to die,
but it failed to actually send the kill signal to it, thanks to a missing -f
option.  This was hard to detect since the babysitter normally dies of its
own free will anyway.

Change-Id: I184f5ae139d85c178f3eb7795fc7d0605aad9ff5