Clean up sh syntax, and don't hardcode model numbers.
Model numbers evolve over time and we shouldn't be hardcoding them in so
many places; instead, use the presence or absence of various packages as an
indicator of which code path to take.

While we're here, clean up a bunch of sh syntax to make it more consistent
with the rest o the boot scripts.  There's probably still more work to be
done on this front.

Change-Id: I2fb82040028708112576e3b7ca795ec1517831b7