gfrg200: enable PACKAGE_GOOGLE_STORAGE_BOX and ginstall images.
This config is largely based on the kvm_defconfig one. It's definitely not
perfect, but it does build all the SageTV dependencies, although it still
uses the wrong jvm.

The ginstall image is not installable until we also teach ginstall how
to deal with it, but the produced uImage and squashfs do work, if loaded
correctly into flash using u-boot.

Change-Id: I1d2875f8237ac7bdbbf75595b1a70b4e9a74af4b