S40network: copy default routes from ip= kernel command line.
When the kernel's ethernet driver is built in, the kernel ip autoconfig can
handle setting IP address and default route.  When it isn't, we need to set
the IP address ourselves (which we already were doing in fs/simpleramfs) and
also the default route (which simpleramfs can't currently do because its
route command is broken and it doesn't include the 'ip' command).

Without this, you can mount NFS root, but you can't get an NTP signal so
nothing really happens.

Change-Id: I779829006b21a0c5cc7f3fe0354de3dc6becfb78