psstat: add a "ccputop" measurement for cumulative (or child) cpu time.
The 'cputop' display only showed the CPU time used by currently-running
processes.  This can be misleading in the case of processes that do a lot of
short-lived fork-exec, since the time used by child processes wasn't being
accounted to anyone.  This is much more realistic.

We could consider just removing the old 'cputop' field (since it's
comparatively not very useful), but let's keep it for now in order not to
upset comparison with older versions.  There's a slight difference that it
will now show bsdtime (format: mmm:ss) instead of cputime (format:
hh:mm:ss), but the meaning is still the same as before.  I changed it to
make it easier to look at the output and compare cputop to ccputop.

Change-Id: Id5280fc7012943806a74f50af8e737856dcb64b0