If a prebuilt u-boot exists, copy that to images.
If a prebuilt binary exists for the current target
(gflt110 for example) copy the prebuilt binary to
u-boot-spi.bin and the built binary to u-boot-spi.bin.built.

This means our prebuilt binary will end up in the ginstall file,
and the built binary is available for testing.

If the prebuilt binary does not exist, then this falls back to the
old behavior of copying the built u-boot file and putting that into
the ginstall archive.

Change-Id: I97f97bf3415966dc8bd0a32c39d075faf3529232